Update how rule usage is recorded
The observer recording rule usage was wrapped in a check for a zero discount and skipped registering rule usage in that case.
Some third party rules result in zero amount discount but secondary results.
The zero value check is pointless as if a rule is applied then it should be recorded in any case.

Additionally, recording rule usage was subject to race conditions. Teh code read the current usage count then saved it back with the value incremented.
Two customers triggering this code at the same time would result in a race since it's no an atomic operation.
This has been replaced with atomic updates that uses DB expressions to increment the usage counters.